Refactor SCSS structure and improve token creation styles
- Create/Improve abstract SCSS files for variables, mixins, extends, and forms
- Simplify and standardize _tokenCreation.scss and _tokenPreview.scss
- Implement consistent use of variables and mixins across styles
- Improve code maintainability and reusability
- Enhance responsiveness with standardized gap and padding values
