+++ src/style/main.scss
@@ -12,7 +12,7 @@
 @import './abstract/typography.scss';
 @import './abstract/shadows.scss';
 @import './abstract/forms.scss';
-
+@import './abstract/extends.scss';
 
 // Base imports
 @import './base/global.scss';
