+++ src/style/pages/tokenCreation/_tokenCreation.scss
@@ -1,163 +1,90 @@
 // src/style/pages/tokenCreation/_tokenCreation.scss
 
 .token-creation {
-  @include flex-center;
-  flex-direction: column;
-  padding: pxToRem(20);
+  @include flex-container(row);
+  gap: $standard-gap;
+  padding: $standard-padding;
 
-  &__form {
+  &__form,
+  &__preview {
     flex: 1;
-    max-width: 50%;
+    max-width: $section-max-width;
   }
 
   &__preview {
-    flex: 1;
-    position: sticky;
-    top: pxToRem(20);
-    max-height: calc(100vh - #{pxToRem(40)});
+    @include sticky-element($standard-padding);
+    align-self: flex-start;
+    max-height: calc(100vh - #{2 * $standard-padding});
     overflow-y: auto;
   }
 
-  h1, h2 {
-    color: var(--primary-color);
-    margin-bottom: pxToRem(20);
-  }
-
   .token-basics {
-    width: 100%;
-    max-width: pxToRem(600);
-    margin-bottom: pxToRem(30);
+    @extend %full-width;
+    margin-bottom: $gap-large;
 
     input, textarea {
-      width: 100%;
-      padding: pxToRem(10);
-      margin-bottom: pxToRem(10);
-      border: 1px solid var(--border-color);
-      border-radius: pxToRem(4);
-      font-size: pxToRem(16);
-      color: var(--font-color);
-      background-color: var(--background-color);
-
-      &:focus {
-        outline: none;
-        border-color: var(--primary-color);
-      }
+      @extend %form-field;
     }
 
     textarea {
-      height: pxToRem(100);
-      resize: vertical;
+      @extend %textarea-field;
     }
   }
 
   .form-builder {
-    width: 100%;
-    max-width: pxToRem(600);
+    @extend %full-width;
 
     .field-buttons {
-      display: flex;
+      @include flex-container;
       justify-content: space-between;
-      margin-bottom: pxToRem(20);
+      margin-bottom: $standard-gap;
 
       button {
-        padding: pxToRem(10) pxToRem(15);
-        background-color: var(--secondary-color);
-        color: var(--button-text-color);
-        border: none;
-        border-radius: pxToRem(4);
-        cursor: pointer;
-        transition: background-color 0.3s ease;
-
-        &:hover {
-          filter: brightness(0.9);
-        }
+        @extend %button-base;
       }
     }
 
     .form-field {
-      background-color: var(--card-color);
-      padding: pxToRem(15);
-      margin-bottom: pxToRem(10);
-      border-radius: pxToRem(4);
-      @extend .shadow__general;
-
-      display: flex;
+      @extend %card-styles;
+      @include flex-container;
       flex-wrap: wrap;
       align-items: center;
-      gap: pxToRem(10);
+      gap: $gap-small;
+      margin-bottom: $gap-small;
 
       input, select {
-        padding: pxToRem(5);
-        border: 1px solid var(--border-color);
-        border-radius: pxToRem(4);
-        font-size: pxToRem(14);
+        @extend %form-field;
+        padding: $input-padding / 2;
+        font-size: $input-font-size * 0.875;
       }
 
       label {
         display: flex;
         align-items: center;
-        gap: pxToRem(5);
+        gap: $gap-small / 2;
       }
 
       button {
-        padding: pxToRem(5) pxToRem(10);
-        background-color: var(--error-color);
-        color: white;
-        border: none;
-        border-radius: pxToRem(4);
-        cursor: pointer;
-        transition: background-color 0.3s ease;
-
-        &:hover {
-          filter: brightness(0.9);
-        }
+        @extend %error-button;
       }
-    }
-  }
-
-  .save-button {
-    margin-top: pxToRem(20);
-    padding: pxToRem(12) pxToRem(20);
-    background-color: var(--button-primary-color);
-    color: var(--button-text-color);
-    border: none;
-    border-radius: pxToRem(4);
-    font-size: pxToRem(16);
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-
-    &:hover {
-      filter: brightness(0.8);
     }
   }
 
   .preview-controls {
-    display: flex;
+    @extend %full-width;
+    @include flex-container;
     justify-content: space-between;
-    margin-top: pxToRem(20);
-    width: 100%;
-    max-width: pxToRem(600);
+    margin-top: $standard-gap;
 
     button {
-      padding: pxToRem(12) pxToRem(20);
-      border: none;
-      border-radius: pxToRem(4);
-      font-size: pxToRem(16);
-      cursor: pointer;
-      transition: background-color 0.3s ease;
+      @extend %button-base;
 
       &.preview-button {
         background-color: var(--secondary-color);
-        color: var(--button-text-color);
       }
 
       &.save-button {
         background-color: var(--button-primary-color);
-        color: var(--button-text-color);
-      }
-
-      &:hover {
-        filter: brightness(0.8);
       }
     }
   }
@@ -175,74 +102,13 @@
       max-height: none;
     }
   }
-
 }
-
-.token-preview {
-  margin-top: pxToRem(30);
-  width: 100%;
-  max-width: pxToRem(600);
-  background-color: var(--card-color);
-  padding: pxToRem(20);
-  border-radius: pxToRem(8);
-  @extend .shadow__general;
-
-  h2 {
-    color: var(--primary-color);
-    margin-bottom: pxToRem(15);
-  }
-
-  h3 {
-    font-size: pxToRem(20);
-    margin-bottom: pxToRem(10);
-  }
-
-  p {
-    margin-bottom: pxToRem(20);
-  }
-
-  .preview-field {
-    margin-bottom: pxToRem(15);
-
-    label {
-      display: block;
-      margin-bottom: pxToRem(5);
-      font-weight: bold;
-    }
-
-    input, select {
-      width: 100%;
-      padding: pxToRem(8);
-      border: 1px solid var(--border-color);
-      border-radius: pxToRem(4);
-      font-size: pxToRem(14);
-    }
-  }
-
-  .preview-submit {
-    width: 100%;
-    padding: pxToRem(10);
-    background-color: var(--button-primary-color);
-    color: var(--button-text-color);
-    border: none;
-    border-radius: pxToRem(4);
-    font-size: pxToRem(16);
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-
-    &:hover {
-      filter: brightness(0.8);
-    }
-  }
-}
-
 
 @include responsive(mobile) {
   .token-creation {
-
     .preview-controls {
       flex-direction: column;
-      gap: pxToRem(10);
+      gap: $gap-small;
 
       button {
         width: 100%;
@@ -252,10 +118,10 @@
     .form-builder {
       .field-buttons {
         flex-wrap: wrap;
-        gap: pxToRem(10);
+        gap: $gap-small;
 
         button {
-          flex: 1 1 calc(50% - #{pxToRem(5)});
+          flex: 1 1 calc(50% - #{$gap-small / 2});
         }
       }
     }
