+++ src/style/pages/tokenCreation/_tokenPreview.scss
@@ -1,79 +1,58 @@
 // src/style/components/tokenCreation/_tokenPreview.scss
 
-// ! Needs colors fixed. The colors are not the same as the design.
-
 .token-preview {
-  background: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
-  border-radius: pxToRem(12);
+  @extend %card-styles;
+  margin-top: $gap-large;
+  width: 100%;
+  max-width: pxToRem(600);
   overflow: hidden;
-  box-shadow: 0 pxToRem(10) pxToRem(20) rgba(0, 0, 0, 0.1);
 
   &__header {
-    background: linear-gradient(to right, #4facfe 0%, #00f2fe 100%);
-    padding: pxToRem(20);
-    color: white;
+    background: linear-gradient(to right, var(--primary-color) 0%, var(--secondary-color) 100%);
+    padding: $standard-padding;
+    color: var(--button-text-color); 
 
     h2 {
       margin: 0;
-      font-size: pxToRem(24);
+      font-size: $heading-font-size;
       font-weight: bold;
     }
   }
 
   &__content {
-    padding: pxToRem(20);
+    padding: $standard-padding;
+  }
 
-    h3 {
-      color: #333;
-      font-size: pxToRem(20);
-      margin-bottom: pxToRem(10);
-    }
+  h3 {
+    font-size: $subheading-font-size;
+    margin-bottom: $gap-small;
+    color: var(--font-color);
+  }
 
-    p {
-      color: #666;
-      margin-bottom: pxToRem(20);
-    }
+  p {
+    margin-bottom: $standard-gap;
+    color: var(--font-color);
   }
 
   .preview-field {
-    margin-bottom: pxToRem(15);
+    margin-bottom: $standard-gap;
 
     label {
       display: block;
-      margin-bottom: pxToRem(5);
+      margin-bottom: $gap-small / 2;
       font-weight: bold;
-      color: #444;
+      color: var(--font-color);
     }
 
     input, select {
-      width: 100%;
-      padding: pxToRem(10);
-      border: 1px solid #ddd;
-      border-radius: pxToRem(4);
-      font-size: pxToRem(14);
-      transition: border-color 0.3s ease;
-
-      &:focus {
-        border-color: #4facfe;
-        outline: none;
-      }
+      @extend %form-field;
     }
   }
 
   .preview-submit {
+    @extend %button-base;
     width: 100%;
-    padding: pxToRem(12);
-    background: linear-gradient(to right, #4facfe 0%, #00f2fe 100%);
-    color: white;
-    border: none;
-    border-radius: pxToRem(4);
-    font-size: pxToRem(16);
     font-weight: bold;
-    cursor: pointer;
-    transition: opacity 0.3s ease;
-
-    &:hover {
-      opacity: 0.9;
-    }
+    background-color: var(--button-primary-color);
   }
 }
